# services/core-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Core Gateway API
  version: 0.1.0
  description: OpenAPI-first scaffold for Phase 1 gateway/auth.
servers:
  - url: http://localhost:8080
paths:
  /modules:
    get:
      operationId: listModules
      summary: List the AscendOS module catalog
      responses:
        '200':
          description: Catalog response
          content:
            application/json:
              schema:
                type: object
                properties:
                  modules:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModuleCatalogEntry'
  /modules/{name}:
    get:
      operationId: getModule
      summary: Fetch a single module by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleCatalogEntry'
        '404':
          description: Module not found
  /healthz:
    get:
      operationId: healthz
      summary: Liveness check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /readyz:
    get:
      operationId: readyz
      summary: Readiness check
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
  /metrics:
    get:
      operationId: metrics
      summary: Prometheus metrics placeholder
      responses:
        '200':
          description: Placeholder metrics response
          content:
            text/plain:
              schema:
                type: string
  /me:
    get:
      operationId: me
      summary: Return validated claims from bearer token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Claims extracted
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Missing or invalid bearer token
components:
  schemas:
    ModuleCatalogEntry:
      type: object
      properties:
        name:
          type: string
        key:
          type: string
        language:
          type: string
        domain:
          type: string
        summary:
          type: string
        servicePath:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT